Update Requirement and Updater

+++ update.sh
@@ -13,6 +13,7 @@ if [[ $ANS == "N" ]]; then
     echo "Good bye, have a good day."
     exit 1
 fi
+ANS=''
 
 echo -e "\n=========================="
 echo "PULL LATEST UPDATES"
@@ -25,8 +26,20 @@ echo -e "\n=========================="
 echo "STOPPING CURRENT PROCESS"
 echo "=========================="
 docker-compose down
-docker volume rm gerobug_static-content
-sleep 3
+
+while [[ ! $ANS =~ ^$rx$ ]]; do
+    echo -e "\nUPDATE STATIC FILES?"
+    echo "If Y, any updates on static files will be applied, but you need to re-upload your logo and templates after the update"
+    echo "If N, Updates on static files will NOT be applied"
+    echo "(Y/N)"
+    read ANS
+    echo
+done
+
+if [[ $ANS == "Y" ]]; then
+    docker volume rm gerobug_static-content
+    sleep 3
+fi
 
 echo -e "\n=========================="
 echo "REBUILD UPDATES"
